Generate metadata about genbank sequences using --inputfile instead of --package
An earlier version of this code worked when using the dataformat command to
generate metadata directly from the zipped package downlaoded via the datasets cli

A few days ago, that stopped working. It's unclear whether this was a transient
error or something more permanent. Regardless, passing the data_report.jsonl file directly
to dataformat (instead of passing the .zipped package via the --package parameter)
fixed the problem. The code already unzips the downloaded package, so the
data_report.jsonl is readily available.