Refactor code that pulls reference tree and wrap code in a CLI (#5)
* Remove PDM

We haven't made an official decision on what will replace it, but the
combination of conda and pdm was causing some weirdness with virtual
environments, so let's switch it out to managing dependencies via
pip-tools, even if only temporarily.

* Generate metadata about genbank sequences using --inputfile instead of --package

An earlier version of this code worked when using the dataformat command to
generate metadata directly from the zipped package downlaoded via the datasets cli

A few days ago, that stopped working. It's unclear whether this was a transient
error or something more permanent. Regardless, passing the data_report.jsonl file directly
to dataformat (instead of passing the .zipped package via the --package parameter)
fixed the problem. The code already unzips the downloaded package, so the
data_report.jsonl is readily available.

* Add dependencies and replace logger cruft with loguru

* Add function to retrieve point-in-time reference trees

Also switch the logging package again

* Extract root_sequence from the reference data returned via nextclade API

For now, let's save a copy of the root_sequence data that comes with
the reference tree information. We believe the root_sequence
information won't change over time

* Put a little CLI on it

* This backfires every time

* Unbork the README

* Add a test harness and a few tests

Also refactor the request reference data function to be more
unit test friendly.