Use released_after instead of updated_after when downloading genome package (#15)
* don't commit intermediate pipeline outputs

* Incorporate timestamp of pipeline run into sequence data filenames

fix type check

* Move tests to the correct location

* move reference sequence temporarily

* Add requirements for merging files

pyarrow is a requirement for the polars to_pandas method
move rich from dev dependency to package dependency, since
we're using rich-click to format cli output

* fixing install instructions again

* Merge clade assignments with sequence metadata

* only commit final assignment files

* Switch from updated_since to released_since when downloading NCBI genome package

Resolves #3

The NCBI links for accessing cached data resulted in a ~20 min
download. We might want that in production, since this pipeline will
run infrequently as a batch job. For our MVP, however, let's stick
to shorter run times using the released_since parameter.

This changeset also switches from using the datasets CLI to NCBI's
API, removing datasets as a project dependency)

* Add retry to NCBI API call

* Rename as_of_date to to make it clearer we're talking about the reference tree

* Improve handling of API request

* Add a CLI option for "released since date"

It's not clear what a "smart" default looks like here, so for
now this is a required parameter when running the assign_clades
process.

* Cleanup data sub-folders and miscellaneous files

Segregating reference/assignment/sequence data in a specific run made it
harder to view the intermediate outputs, so getting ready for a change
to for a change to group all of a run's files in a single directory

* Replace the global variables with a config object

Take one more step towards a more reasonable-looking code base. The config
isn't user-updatable, but it's now isolated into a single file for easier
updates. Additionally, put all intermediate files into a single directory
instead of spreading them across three different folders.

* Add coverage package

It's not yet hooked up to github, but the coverage report is
useful for local development

* Add warning about not using ~ when supplying a data path

* Remove streaming/iter_chunk when saving NCBI .zip package

* Switch dev-requirements to requirements-dev for consistency

* Switch to python zip module to unzip NCBI package