Use root sequence returned by nextstrain API
This changeset replaces the static COVID root sequence that was
stored in the repo with the root sequence returned as part of
the nextrain API call.

Note that the root sequence returned by the API call doesn't contain
a fasta header, so the code manually adds one (the nextstrain cli
that assigns sequences will fail if the specified root sequence
file isn't in fasta format). Manipulating the data this way
doesn't feel great, but in the spirit of using data from a
canonical source instead of storing our copies, this seems
preferable.

remove the root sequence file