Use sequence metadata as of the round closing time rather than currnet time
This change will make the script more precise and more accurate
should we need to re-run it. Rather than using the version of
genome sequence metadata at the time script is run, this
update explicitly uses the sequence metadata as it existed at
the time the most recent round closed.

The update generated a slightly different version of the resulting
data, which is included in the commit.