Generate a list of sequence collections by date and location (#92)
* Add a script that counts sequences reported by location and date

By default, the dates included will be 31 days prior to the date
of the round that just closed (the script will be scheduled to
run on Wednesday evenings)

* Use sequence metadata as of the round closing time rather than currnet time

This change will make the script more precise and more accurate
should we need to re-run it. Rather than using the version of
genome sequence metadata at the time script is run, this
update explicitly uses the sequence metadata as it existed at
the time the most recent round closed.

The update generated a slightly different version of the resulting
data, which is included in the commit.

* Add a check to the computed location/state counts

* Add some logging, as a treat

* Change virus-clade-utils back to main branch

* Enable streaming mode on the Polars collect() operation

Although the documentation considers streaming to be "unstable",
this changeset adds it as a hedge against out of memory errors.

* Add GitHub action to run after round close

Introduce another GitHub workflow that runs on Wednesdays after
a rounds closes. It generates a count of Sars-Cov-2 sequences
collected for the past 31 days (grouped by location and day).

* GitHub workflow cleanup

uv is a drop-in replacement for the setup-python action,
so we don't need the latter

* Eliminate date pivot from the output .csv

* rename folder for summarized location/date counts

* Add location of unscored location/date file to README

* Update for changes in the virus-clade-utils repo

The virus-clade-utils repo is now named cladetime. The
virus_clade_utils Python package is now named cladetime.
More information:
https://github.com/reichlab/cladetime/issues/31