Move all "clades to model" logic to the hub
A prior verion of get_clades_to_model.py was invoking a
get_clade_list.py script in the cladetime package. Now that
the CladeTime interface makes it easier to work with
sequence metadata, we can completely de-couple this
hub-specific logic from cladetime and keep all of the
"get clade list" code with the hub repo, where it logically
belongs.