update renv packages; pin hub pkgs to release
I updated the renv packages this way:

```r
renv::update()
pkgs <- c("hubverse-org/hubValidations@*release",
  "hubverse-org/hubAdmin@*release",
  "hubverse-org/hubData@*release",
  "hubverse-org/hubUtils@*release"
)
renv::install(pkgs)
renv::snapshot()
```

Note that the hubverse packages for the moment need to be manually
updated in the manner presented. There is a proposal for creating an
archived repository in #137, but until that is created, this will have
to work.

This will fix #137