Coerce text columns to Arrow string type
Prior to this change, target data files were created using the Polars
write_parquet method. However, that operation results in column datatypes
that are interpreted by Arrow as string_large. Because we're including our
partition fields in the parquet file, the large_string results in a type
mismatch when R packages use arrow::open_dataset to read the files (the value
in the partition key is read as a string, not a large_string)