Bs/fix target date schema/265 (#267)
* Use Nextstrain 100K files for integration tests

* Coerce text columns to Arrow string type

Prior to this change, target data files were created using the Polars
write_parquet method. However, that operation results in column datatypes
that are interpreted by Arrow as string_large. Because we're including our
partition fields in the parquet file, the large_string results in a type
mismatch when R packages use arrow::open_dataset to read the files (the value
in the partition key is read as a string, not a large_string)

* Match target-output schema to schema used for model-output files

Assuming that any Hubverse tools using arrow::open_dataset() when reading
target-data files, we should be to use the hub's model output schema when
writing them. hubDataPy isn't online yet (which will presumably be able to
determine a hub's model-output schema like it's R counterpart), so this
changeset manually creates a schema to use.

* Update tests for data type changes