Update oracle target data files
This changeset adds an as_of column to oracle target data files. In this
case, as_of represents a sequence_as_of date. It's not strictly necessary
to include in the oracle target data, but we decided to add it as a breadcrumb
in case an oracle output file is overwritten via a manual target data run.