Update target-data files (#343)
* Update timeseries target data files

This changeset updates the hub's target time series files as described
on #296:

- Partition by as_of first (renamed from sequence_as_of) and then by nowcast_date
- Remove the tree_as_of column

* Update oracle target data files

This changeset adds an as_of column to oracle target data files. In this
case, as_of represents a sequence_as_of date. It's not strictly necessary
to include in the oracle target data, but we decided to add it as a breadcrumb
in case an oracle output file is overwritten via a manual target data run.

* Change data type of observation and oracle_value columns

The oracle oracle_value column and the timeseries observation columns
reflect the number of sequences for a specific collection date/location/clade.
This changeset updates those data types from float to int to more accurately
reflect the contents of those columns.

* Add data dictionaries for oracle output and timeseries target-data

* Remove existing target data from variant-nowcast-hub

Delete the existing target-date to ensure we don't have any lingering files
with the old formats before we start backfilling.

* Install cladetime from PyPi instead of git 🎉