Better document the various dates at play when creating target data
This changeset doesn't affect the location or contents of the target
data being created. Instead it attempts to address some of the confusion
around various dates. For example, we now explicitly set a "round_close_date"
(13 weeks/91 days before the nowcast_date). Then, sequence_as_of is
calculated as round_close_date + 90 days (to make the "why" of the
sequence_as_of clearer).

It also improves the title and and body of the PR to make the dates clearer
and emits some additional information while running the post submission jobs
workflow.