Make target data dates, their origin, and their purpose clearer (#363)
* Better document the various dates at play when creating target data

This changeset doesn't affect the location or contents of the target
data being created. Instead it attempts to address some of the confusion
around various dates. For example, we now explicitly set a "round_close_date"
(13 weeks/91 days before the nowcast_date). Then, sequence_as_of is
calculated as round_close_date + 90 days (to make the "why" of the
sequence_as_of clearer).

It also improves the title and and body of the PR to make the dates clearer
and emits some additional information while running the post submission jobs
workflow.

* Log all get_target_data parameters in a single place

* Add more detail to target data README

This changeset adds a section about the various dates used when creating
target data as well as few sentences about the process. It also adds
some acknowledgements, since we're using Nextclade to do our custom clade
assignments.

* Fix PR creation