Consolidate and fix variant-nowcast-hub Python test workflows (#457)
* Update get_location_date_counts test

Instead of running the test whenever get_location_date_counts.py is
executed, create a test that can be run via pytest. This change lays the
foundation for automating the test run when a PR is opened (and matches
the pattern used by the other scripts in the src/ directory)

* Combine Python test workflows

The Python scripts change infrequently enough that the simplicity advantage of
having a single workflow to test all three outweigh the extra test runs.

This changeset also adds an invocation for the get_location_date_counts tests
that were modified in the prior commit.

* Don't get fancy with the working directory

* Rename job for clarity