Remove the streaming option for Polars collect() operation
This is the only script in src/ that uses the Polars streaming option.
It's also the script that deals with the smallest amount of data, so this
commit removes streaming altogether.

Streaming is a great Polars feature, but it's still in development. Thus,
we should avoid it in the hub's operational workflow unless absolutely
necessary.