Address polars streaming shenanigans (#459)
* Update Polars streaming example to use the new engine

Polars has been implementing a new streaming engine:
https://github.com/pola-rs/polars/issues/20947

This changeset updates the Polars streaming syntax in the repo's README
to reflect the new engine. The new engine is still under development (but
so was the deprecated one!)

I tested this change on my laptop (which is a beast). Regardless, as the
Polars team works on the new engine, the older one will become less stable
over time. It's now time to make the swtich (which will also prevent
the "streaming=True" deprecation message)

* Remove the streaming option for Polars collect() operation

This is the only script in src/ that uses the Polars streaming option.
It's also the script that deals with the smallest amount of data, so this
commit removes streaming altogether.

Streaming is a great Polars feature, but it's still in development. Thus,
we should avoid it in the hub's operational workflow unless absolutely
necessary.